Refactor config workflow
