+++ crates/smart-config/src/lib.rs
@@ -1,3 +1,6 @@
 pub mod metadata;
 mod parsing;
+mod source;
 mod value;
+
+pub use self::source::{ConfigRepository, Environment};
